backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
segno==1.6.1
pillow==10.2.0
redis==5.0.1
python-jose==3.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2
httpx==0.27.0
